refactor: move shared files into "shared" folder
