+++ src/configs/jsonc.ts
@@ -1,9 +1,9 @@
-import type { FilesOptions, OverridesOptions, StylisticOptions, TypedConfigArray } from '../types';
+import type { FilesOptions, OverridesOptions, StylisticOptions, TypedConfigArray } from '../shared/types';
 
 import tseslint from 'typescript-eslint';
 
-import { GLOB_JSON, GLOB_JSON5, GLOB_JSONC } from '../globs';
-import { interopDefault } from '../utils';
+import { GLOB_JSON, GLOB_JSON5, GLOB_JSONC } from '../shared/globs';
+import { interopDefault } from '../shared/utils';
 
 /**
  * Generates an ESLint configuration for JSONC files.
