+++ src/index.ts
@@ -2,6 +2,6 @@ export * from './configs';
 export * from './factories/project-config';
 export * from './factories/standard-config';
 export * from './factories/workspace-config';
-export { GLOB_HTML, GLOB_JS, GLOB_SRC, GLOB_TESTS, GLOB_TS } from './globs';
-export type * from './types';
-export * from './utils';
+export { GLOB_HTML, GLOB_JS, GLOB_SRC, GLOB_TESTS, GLOB_TS } from './shared/globs';
+export type * from './shared/types';
+export * from './shared/utils';
